fix: 바우처 필드값 타입 변경

+++ src/main/java/org/prgrms/kdt/domain/voucher/model/FixedAmountVoucher.java
@@ -4,9 +4,9 @@ import java.util.UUID;
 
 public class FixedAmountVoucher implements Voucher{
     private final UUID voucherId;
-    private final int discount;
+    private final long discount;
 
-    public FixedAmountVoucher(UUID voucherId, int discount) {
+    public FixedAmountVoucher(UUID voucherId, long discount) {
         this.voucherId = voucherId;
         this.discount = discount;
     }
